Wait for rendering to complete at the end
This is not needed, as this will happen anyway when the texture is
read into CPU memory. However, this differentates between time taken
to save the image and time taken to render the image. Saving the image
should be fairly quick, whereas finishing rendering may take a long
time if many commands have been queued.