Fix texture and framebuffer initialisation and binding
Textures were not being bound correctly, and the framebuffer was using
a datatype that did not provide sufficient accuracy for the error
function, and complicated things due to clamping. The datatype now
defaults to RGBA, but for the error function we specify that it should
use RGBA32F, which seems to give a better accuracy.