Debugging code to test the contents of the difference texture
For smaller errors, the result is not quite correct when compared with
results predicted by numpy. This is thought to be due to the floating
point accuracy. Using RGBA32F has improved this, but it is still not
completely correct. This can be investigated further. However, larger
errors match the predictions of numpy more or less exactly.